refactor: Separate promptUser into a common module
