+++ src/view/ui/common/input.js
@@ -1,0 +1,15 @@
+import readline from "readline";
+
+export function promptUser(query) {
+  return new Promise((resolve) => {
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(query, (input) => {
+      rl.close();
+      resolve(input.trim());
+    });
+  });
+}
